requirements.txt
streamlit==0.74.1
albumentations==0.5.2
opencv-python==4.5.1.48
Pillow==8.1.0
torchvision==0.8.2